requirements.txt
paramiko
pyyaml
flask
pyte
requests